chore(examples/multi-tenant): migrates to 2.0
