+++ examples/multi-tenant/.env.example
@@ -1,5 +1,5 @@
-MONGODB_URI=mongodb://127.0.0.1/payload-example-auth
-PAYLOAD_SECRET=PAYLOAD_AUTH_EXAMPLE_SECRET_KEY
+DATABASE_URI=mongodb://127.0.0.1/payload-example-multi-tenant
+PAYLOAD_SECRET=PAYLOAD_MULTI_TENANT_EXAMPLE_SECRET_KEY
 PAYLOAD_PUBLIC_SERVER_URL=http://localhost:3000
 PAYLOAD_SEED=true
 PAYLOAD_DROP_DATABASE=true
