+++ examples/multi-tenant/src/access/anyone.ts
