+++ examples/multi-tenant/src/collections/Pages/index.ts
@@ -1,7 +1,7 @@
 import type { CollectionConfig } from 'payload/types'
 
-import richText from '../fields/richText'
-import { tenant } from '../fields/tenant'
+import richText from '../../fields/richText'
+import { tenant } from '../../fields/tenant'
 import { loggedIn } from './access/loggedIn'
 import { tenantAdmins } from './access/tenantAdmins'
 import { tenants } from './access/tenants'
