+++ examples/multi-tenant/src/collections/Tenants/index.ts
@@ -1,6 +1,6 @@
 import type { CollectionConfig } from 'payload/types'
 
-import { superAdmins } from '../access/superAdmins'
+import { superAdmins } from '../../access/superAdmins'
 import { tenantAdmins } from './access/tenantAdmins'
 
 export const Tenants: CollectionConfig = {
