+++ examples/multi-tenant/src/collections/Users/access/adminsAndSelf.ts
@@ -1,7 +1,7 @@
 import type { Access } from 'payload/config'
-import type { User } from 'payload/generated-types'
 
-import { isSuperAdmin } from '../../utilities/isSuperAdmin'
+import { isSuperAdmin } from '../../../utilities/isSuperAdmin'
+import { User } from '../../../payload-types'
 
 export const adminsAndSelf: Access<any, User> = async ({ req: { user } }) => {
   if (user) {
