+++ examples/multi-tenant/src/collections/Users/access/tenantAdmins.ts
@@ -1,6 +1,6 @@
 import type { FieldAccess } from 'payload/types'
 
-import { checkUserRoles } from '../../utilities/checkUserRoles'
+import { checkUserRoles } from '../../../utilities/checkUserRoles'
 import { checkTenantRoles } from '../utilities/checkTenantRoles'
 
 export const tenantAdmins: FieldAccess = args => {
