--- examples/multi-tenant/src/collections/fields/richText/elements.ts
@@ -1,5 +1,0 @@
-import type { RichTextElement } from 'payload/dist/fields/config/types'
-
-const elements: RichTextElement[] = ['blockquote', 'h2', 'h3', 'h4', 'h5', 'h6', 'link']
-
-export default elements
