+++ examples/multi-tenant/src/dotenv.js
@@ -1,0 +1,3 @@
+module.exports = {
+  config: () => null,
+}
