+++ examples/multi-tenant/src/fields/richText/elements.ts
@@ -1,0 +1,5 @@
+import type { RichTextElement } from '@payloadcms/richtext-slate'
+
+const elements: RichTextElement[] = ['h1', 'h2', 'h3', 'h4', 'h5', 'h6', 'blockquote', 'link']
+
+export default elements
