+++ examples/multi-tenant/src/payload-types.ts
@@ -1,4 +1,5 @@
 /* tslint:disable */
+/* eslint-disable */
 /**
  * This file was automatically generated by Payload.
  * DO NOT MODIFY IT BY HAND. Instead, modify your source Payload config,
@@ -10,6 +11,8 @@ export interface Config {
     users: User
     tenants: Tenant
     pages: Page
+    'payload-preferences': PayloadPreference
+    'payload-migrations': PayloadMigration
   }
   globals: {}
 }
@@ -17,18 +20,20 @@ export interface User {
   id: string
   firstName?: string
   lastName?: string
-  roles: Array<'super-admin' | 'user'>
-  tenants?: Array<{
+  roles: ('super-admin' | 'user')[]
+  tenants?: {
     tenant: string | Tenant
-    roles: Array<'admin' | 'user'>
+    roles: ('admin' | 'user')[]
     id?: string
-  }>
+  }[]
   lastLoggedInTenant?: string | Tenant
   updatedAt: string
   createdAt: string
-  email?: string
+  email: string
   resetPasswordToken?: string
   resetPasswordExpiration?: string
+  salt?: string
+  hash?: string
   loginAttempts?: number
   lockUntil?: string
   password?: string
@@ -36,10 +41,10 @@ export interface User {
 export interface Tenant {
   id: string
   name: string
-  domains: Array<{
+  domains?: {
     domain: string
     id?: string
-  }>
+  }[]
   updatedAt: string
   createdAt: string
 }
@@ -48,9 +53,47 @@ export interface Page {
   title: string
   slug?: string
   tenant?: string | Tenant
-  richText: Array<{
+  richText: {
     [k: string]: unknown
-  }>
+  }[]
   updatedAt: string
   createdAt: string
+}
+export interface PayloadPreference {
+  id: string
+  user: {
+    relationTo: 'users'
+    value: string | User
+  }
+  key?: string
+  value?:
+    | {
+        [k: string]: unknown
+      }
+    | unknown[]
+    | string
+    | number
+    | boolean
+    | null
+  updatedAt: string
+  createdAt: string
+}
+export interface PayloadMigration {
+  id: string
+  name?: string
+  batch?: number
+  updatedAt: string
+  createdAt: string
+}
+
+declare module 'payload' {
+  export interface GeneratedTypes {
+    collections: {
+      users: User
+      tenants: Tenant
+      pages: Page
+      'payload-preferences': PayloadPreference
+      'payload-migrations': PayloadMigration
+    }
+  }
 }
