+++ examples/multi-tenant/src/payload.config.ts
@@ -1,3 +1,6 @@
+import { webpackBundler } from '@payloadcms/bundler-webpack'
+import { mongooseAdapter } from '@payloadcms/db-mongodb'
+import { slateEditor } from '@payloadcms/richtext-slate'
 import dotenv from 'dotenv'
 import path from 'path'
 
@@ -13,6 +16,23 @@ import { Users } from './collections/Users'
 
 export default buildConfig({
   collections: [Users, Tenants, Pages],
+  admin: {
+    bundler: webpackBundler(),
+    webpack: config => ({
+      ...config,
+      resolve: {
+        ...config.resolve,
+        alias: {
+          ...config.resolve.alias,
+          dotenv: path.resolve(__dirname, './dotenv.js'),
+        },
+      },
+    }),
+  },
+  editor: slateEditor({}),
+  db: mongooseAdapter({
+    url: process.env.DATABASE_URI,
+  }),
   typescript: {
     outputFile: path.resolve(__dirname, 'payload-types.ts'),
   },
