+++ examples/multi-tenant/src/utilities/isSuperAdmin.ts
@@ -1,5 +1,4 @@
-import type { User } from 'payload/generated-types'
-
+import { User } from '../payload-types'
 import { checkUserRoles } from './checkUserRoles'
 
 export const isSuperAdmin = (user: User): boolean => checkUserRoles(['super-admin'], user)
