+++ examples/multi-tenant/tsconfig.json
@@ -16,7 +16,6 @@
     "sourceMap": true,
     "resolveJsonModule": true,
     "paths": {
-      "payload/generated-types": ["./src/payload-types.ts"],
       "node_modules/*": ["./node_modules/*"]
     },
   },
